Login and signup
The model is being uploaded correctly to the database, we still have to figure out how to handle when someone either logs in or signs up, but for now we can debug using ResponseEntity. We will focus on security eventually, as for now we are uploading plain text passwords to the database.

Co-Authored-By: Andreas4122002 <79724220+Andreas4122002@users.noreply.github.com>