Purchase history Controller and Mustache
Test is pending, Mustache and the controller have been updated accordingly. Now the system should grab the ids from the URLs, so the review and cancel buttons should work.
The purchase model now allows only for one Product, easier access when we try to display and manage them.
They should work or require minimal changes once we add products to the database to test.