Automatic confirmation email
Is working, it is hideous and I have to figure out how to get Mustache to work but it works for testing purposes. The Outlook account is not flagged, but could easily get flagged as spam so we have to figure out how to handle that, Google doesn't allow non-secure apps to access SMTP anymore. And having the public repo is a risk, as our password is shared with the internet on the properties and the MailSender Bean configuration :+1: