Product creation image uploads
Now allows for a main image and multiple images.
The main image is referenced on the product's table, while the list of images is referenced from a separate table called product_images. IDs match and the blobs are being correctly uploaded.
We should change the database structure though, many of the Images parameters are not needed as they are usually referenced by a foreign key.