Added headers and pageable (AJAX JSON)
Now we have a header controller for every page, handling a different set of buttons depending on the level of authority.
The Ajax button now calls for more results and receives the data as JSON (the controller creates a hashmap and converts it to JSON), parses it and builds the new HTML for each product, then appends it to the results div. We consider it is a better approach than using StringBuilder (what we used before).